pyproject.toml
[tool.poetry]
name = "qcc-gateway-hl7-listener"
version = "1.0.0"
description = "HL7 listener service for QCC gateway appliance"
authors = ["sajal-g <sajalgoyal@gmail.com>"]

packages = [
    { include = "hl7_listener", from = "src/main/py"},    
]

[[tool.poetry.source]]
name = "development-pypi"
url = "https://coverahealth.jfrog.io/artifactory/api/pypi/development-pypi/simple"
priority = "primary"

[[tool.poetry.source]]
name = "pypi-public"
url = "https://pypi.org/simple/"
priority = "supplemental"

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.8.3"
hl7 = "^0.4.5"
pydantic = "^1.10.7"
nats-py = "^2.4.0"
covera-logging = "^1.0.9"
covera-tracing = "^1.0.1"
covera-cloud-integration = "^1.0.30"

[tool.poetry.group.test.dependencies]
pytest = "7.1.2"
pytest-mock = "3.8.2"
pytest-asyncio = "0.19.0"
pytest-cov = "^4.1.0"

[tool.poetry.group.dev.dependencies]
docformatter = "^1.5.1"
isort = "^5.12.0"
black = "^23.9.1"

[tool.pytest.ini_options]
testpaths = [
    "src/test",
]

[tool.docformatter]
recursive = true
wrap-descriptions = 88
wrap-summaries = 88

[tool.isort]
profile = "black"
combine_as_imports = true
combine_star = true
multi_line_output = 3
force_grid_wrap = 2
lines_after_imports = 2
sections = "FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"

[tool.black]
target-versions = ["py37", "py38", "py39", "py310"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
